requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.8.0
jupyter-core==5.9.1
kiwisolver==1.5.0
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
onnxruntime==1.24.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.4
polars==1.39.3
polars-runtime-32==1.39.3
prompt-toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
pure-eval==0.2.3
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
scipy==1.17.1
setuptools==81.0.0
simsimd==6.5.16
six==1.17.0
stack-data==0.6.3
stringzilla==4.6.0
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.2
ultralytics==8.4.26
ultralytics-thop==2.0.18
urllib3==2.6.3
wcwidth==0.6.0
python-dotenv
